Simplify CLI
